Deprecate old parser and move to core
